Move logger from log.ts to configurable.ts
